# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/carriers": {
      "get": {
        "description": "Liste les connecteurs REELLEMENT implementes, avec les champs d identifiants qu ils attendent. Un transporteur planifie mais non implemente n y figure pas : le produit ne promet que ce qu il tient.",
        "operationId": "ShipmentsController_listCarriers_v1",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connecteurs transporteurs disponibles",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/carrier-accounts": {
      "get": {
        "operationId": "ShipmentsController_listAccounts_v1",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comptes transporteur configures pour la boutique",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
...
    "/api/v1/orders/{orderId}/ship": {
      "post": {
        "description": "Cree le colis chez le transporteur et passe la commande a EXPEDIEE. Un appel rejoue retourne le colis existant au lieu d en creer un second : trois barrieres independantes l en empechent (V2 §17).",
        "operationId": "ShipmentsController_ship_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "description": "Cle d idempotence. Un rejeu avec la meme cle et le meme corps renvoie la reponse de la premiere execution, sans creer de doublon.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confier une commande au transporteur",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/orders/{orderId}/shipments": {
      "get": {
        "operationId": "ShipmentsController_listByOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Colis d une commande",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/shipments": {
      "get": {
        "description": "Suivi transverse, independant des commandes. Le dernier evenement connu et la date de derniere synchronisation accompagnent chaque colis : un colis sans nouvelle depuis longtemps est un colis a verifier.",
        "operationId": "ShipmentsController_listShipments_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page, 1-indexee.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Taille de page. Plafonnee a 200.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Statuts internes, separes par des virgules.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "DRAFT",
                  "CREATION_PENDING",
                  "CREATED",
                  "PICKED_UP",
                  "IN_TRANSIT",
                  "OUT_FOR_DELIVERY",
                  "DELIVERED",
                  "FAILED_ATTEMPT",
                  "RETURNING",
                  "RETURNED",
                  "CANCELLED",
                  "ERROR"
                ]
              }
            }
          },
          {
            "name": "carrierId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Numero de suivi, reference de commande ou nom du client.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les colis de la boutique",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/shipments/{id}": {
      "get": {
        "operationId": "ShipmentsController_getShipment_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail d un colis et de ses evenements",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/shipments/{id}/cancel": {
      "post": {
        "description": "Annule chez le transporteur puis localement. Un colis deja livre ou retourne ne peut plus etre annule.",
        "operationId": "ShipmentsController_cancelShipment_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelShipmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annuler un colis",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/shipments/{id}/sync-tracking": {
      "post": {
        "description": "Interroge le transporteur et applique les nouveaux evenements. Les evenements deja connus sont ignores : rejouer est sans effet.",
        "operationId": "ShipmentsController_syncTracking_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Forcer une synchronisation du suivi",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/shipments/{id}/tracking": {
      "get": {
        "description": "Conserve le statut BRUT du transporteur et le statut normalise EcomFlow : le premier sert au diagnostic, le second au metier.",
        "operationId": "ShipmentsController_trackingEvents_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Evenements de suivi d un colis",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/returns": {
      "get": {
        "operationId": "ShipmentsController_listReturns_v1",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les retours",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/returns/{id}": {
      "get": {
        "operationId": "ShipmentsController_getReturn_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail d un retour",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/orders/{orderId}/returns": {
      "post": {
        "description": "Le retour reste rattache a sa commande d origine. Aucun mouvement de stock n a lieu ici : la marchandise n est pas encore revenue.",
        "operationId": "ShipmentsController_createReturn_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creer un retour",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/returns/{id}/in-transit": {
      "post": {
        "operationId": "ShipmentsController_markInTransit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marquer le retour comme reparti chez le transporteur",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
    "/api/v1/returns/{id}/received": {
      "post": {
        "operationId": "ShipmentsController_markReceived_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marquer le colis retour comme physiquement recu",
        "tags": [
          "Expedition, suivi et retours"
        ]
      }
    },
...
  },
  "info": {
    "title": "EcomFlow API",
    "description": "API de la plateforme SaaS EcomFlow — gestion et automatisation des operations e-commerce (commandes, confirmation, preparation, expedition, tracking, retours, stock, abonnement).\n\n**Multi-tenant** : chaque requete authentifiee est automatiquement limitee a la boutique de la session. L en-tete `X-Tenant-Id` permet de changer de boutique parmi celles dont l utilisateur est membre ; toute autre valeur est refusee.\n\n**Idempotence** : les endpoints de creation couteuse acceptent l en-tete `Idempotency-Key`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateShipmentDto": {
        "type": "object",
        "properties": {
          "carrierAccountId": {
            "type": "string",
            "description": "Compte transporteur. A defaut, celui par defaut de la boutique."
          },
          "deliveryType": {
            "type": "string",
            "enum": [
              "HOME",
              "PICKUP_POINT"
            ],
            "default": "HOME"
          },
          "pickupPointId": {
            "type": "string",
            "description": "Point relais, requis si deliveryType=PICKUP_POINT."
          },
          "weightGrams": {
            "type": "number",
            "description": "Poids du colis en grammes."
          },
          "notes": {
            "type": "string",
            "description": "Consignes pour le livreur."
          },
          "allowOpening": {
            "type": "boolean",
            "description": "Le client peut-il ouvrir le colis avant paiement ?",
            "default": true
          }
        }
      },
      "CancelShipmentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Motif d annulation, transmis a l historique."
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}